fix(capture): clippy cleanup + Windows Playwright resolver
- raii.rs: add # Safety doc on CVPixelBufferHandle::retain
- sck_backend.rs: drop unnecessary u32 casts (disp.width/height already u32)
- automation.rs: resolve_playwright_target gains Windows branch — calls
  windows::window::find_window_by_pid(pid, None) and wraps the HWND into
  WindowId. With the resolver doing the retry once, downstream target-
  starts use the Window { window_id } path and the WGC backend's
  WindowByPid retry never runs from the command layer.

+++ crates/capture/src/macos/sck_backend.rs
@@ -160,8 +160,8 @@ impl SckBackend {
                     .find(|d| d.display_id() as u64 == display_id.0)
                     .or_else(|| displays.first())
                     .ok_or_else(|| CaptureError::Native("no SCDisplay available".into()))?;
-                let width = disp.width() as u32;
-                let height = disp.height() as u32;
+                let width = disp.width();
+                let height = disp.height();
                 // NEVER pass empty excluding_windows (Pitfall 2) — the
                 // builder default takes care of it when we don't call
                 // `with_excluding_windows`.
